検索用カラムの更新処理もスキーマの取得処理に含める修正 スキーマの取得処理箇所を追加
